reorganized the files for the launch demo
